README.md
This README.md provides an overview of TrekApp, a web platform designed for adventure enthusiasts seeking unforgettable trekking experiences. It outlines the key features, technologies used, installation instructions, and contribution guidelines. 

We have implemented a backend, frontend and a admin dashboard which oversees multiple functionalities such as but not limited to Authentication, Styling, Data Fetching, Deployment and Storage. 

The website is designed to fit your dynamic and responsive needs.